Backend/requirement.txt
# ---------------- Core Backend ----------------
Flask
flask-cors
python-dotenv

# ---------------- Database ----------------
pymongo

# ---------------- OpenAI (New SDK) ----------------
openai

# ---------------- PDF Processing ----------------
pypdf

# ---------------- Vector Search (RAG) ----------------
faiss-cpu
numpy

# ---------------- Utilities ----------------
requests
